+++ tests/test_client.py
@@ -1,12 +1,8 @@
-import platform
+import pytest
 from unittest import TestCase
 
-import pytest
-from unittest import mock
 from alma import ApiModes, Client
 from alma.credentials import ApiKeyCredentials, MerchantIdCredentials, AlmaSessionCredentials
-from alma.request import Request
-from alma.version import __version__ as alma_version
 
 
 class ClientTest(TestCase):
@@ -94,102 +90,3 @@ class ClientTest(TestCase):
 
         # Exposed but empty
         assert alma_client.merchants
-
-
-class ContextTest(TestCase):
-    def setUp(self):
-        client = Client.with_api_key("sk_live_3geNR4OVI0gQGCAKgcYqQs2I")
-        self.context = client.context
-
-    def test_context_expose_logger(self):
-        assert self.context.logger.name == "alma-python-client"
-        assert self.context.logger.level == 0
-
-    def test_context_expose_api_root_and_mode(self):
-        assert self.context.api_root == self.context.options["api_root"]
-        assert self.context.mode == self.context.options["mode"]
-
-    def test_context_expose_url_for(self):
-        assert self.context.url_for("/path/test") == "https://api.getalma.eu/path/test"
-
-    def test_context_expose_user_agent_string(self):
-        assert self.context.user_agent_string().startswith(
-            "alma-client/{alma_version}; Python/{python_version}".format(
-                alma_version=alma_version, python_version=platform.python_version()
-            )
-        )
-
-    def test_context_exposes_credentials(self):
-        assert self.context.credentials
-
-
-class RequestTest(TestCase):
-    def setUp(self) -> None:
-        client = Client.with_api_key("sk_test_3geNR4OVI0gQGCAKgcYqQs2I")
-        self.credentials = client.context.credentials
-        self.request = Request(client.context, "https://url.com")
-
-    def assert_method_calls_configure(self, method):
-        with mock.patch(
-            "alma.credentials.ApiKeyCredentials.configure"
-        ) as mocked_configure, mock.patch("alma.request.requests"):
-            getattr(self.request, method)()
-            mocked_configure.assert_called_once_with(self.request)
-
-    def test_credentials_configure_is_called_on_get(self):
-        self.assert_method_calls_configure("get")
-
-    def test_credentials_configure_is_called_on_post(self):
-        self.assert_method_calls_configure("post")
-
-    def test_credentials_configure_is_called_on_put(self):
-        self.assert_method_calls_configure("put")
-
-
-class ApiKeyCredentialsTest(TestCase):
-    def setUp(self) -> None:
-        self.api_key = "sk_test_3geNR4OVI0gQGCAKgcYqQs2I"
-        client = Client.with_api_key(self.api_key)
-
-        self.credentials = client.context.credentials
-        self.request = Request(client.context, "https://url.com")
-
-    def test_it_adds_authentication_header(self):
-        self.credentials.configure(self.request)
-        assert "Authorization" in self.request.headers
-        assert self.request.headers["Authorization"] == "Alma-Auth {api_key}".format(
-            api_key=self.api_key
-        )
-
-
-class MerchantIdCredentialsTest(TestCase):
-    def setUp(self) -> None:
-        self.merchant_id = "merchant"
-        client = Client.with_merchant_id(self.merchant_id)
-
-        self.credentials = client.context.credentials
-        self.request = Request(client.context, "https://url.com")
-
-    def test_it_adds_authentication_header(self):
-        self.credentials.configure(self.request)
-
-        assert "Authorization" in self.request.headers
-        assert self.request.headers["Authorization"] == "Alma-Merchant-Auth {merchant_id}".format(
-            merchant_id=self.merchant_id
-        )
-
-
-class AlmaSessionCredentialsTest(TestCase):
-    def setUp(self) -> None:
-        self.session_id = "1234567890"
-        self.cookie_name = "alma_session"
-        client = Client.with_alma_session(self.session_id, self.cookie_name)
-
-        self.credentials = client.context.credentials
-        self.request = Request(client.context, "https://url.com")
-
-    def test_it_adds_authentication_header(self):
-        self.credentials.configure(self.request)
-
-        assert self.cookie_name in self.request.cookies
-        assert self.request.cookies[self.cookie_name] == self.session_id
